% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A_thaliana_compress}
\alias{A_thaliana_compress}
\title{A thaliana functional enrichment analysis results for
 "AID","DCE","RCD","SPS" cancer-hallmarks}
\format{
A data frame with 120 rows and 6 variables  (30 GO terms per cancer hallmark):
\describe{
  \item{Enrichment_FDR}{Numeric: False discovery rate values for the GO term}
  \item{Genes_in_list}{numeric: Number of genes in the list of genes for a given GO term}
  \item{Total_genes}{numeric: Number of genes in the genome of a species for a given GO term}
  \item{Functional_Category}{character: GO term name or GO term id}
  \item{Genes}{character: Genes found fot a given GO term}
  \item{feature}{character: A column representing the belonging of a group of comparison}
}
}
\source{
\url{https://data.mendeley.com/datasets/myyy2wxd59/1}
}
\usage{
A_thaliana_compress
}
\description{
This dataset is a subset of the original dataset obtained
 for Clavijo-Buriticá (In preparation)
}
\references{
Clavijo-Buriticá, Sosa, C.C., Mosquera, A.J. Álvarez, A., Medina, J. Quimbaya,
M.A. A systematic comparison of the molecular machinery associated with Cancer-Hallmarks
between plants and humans reveals Arabidopsis thaliana as a useful model to understand
specific carcinogenic events (to be submitted, Target journal: Plos Biology)
}
\keyword{datasets}
