% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareGOspecies.R
\name{compareGOspecies}
\alias{compareGOspecies}
\title{Comprehensive visual and graph comparison between two species and a series of categories}
\usage{
compareGOspecies(df1, df2, GOterm_field, species1, species2)
}
\arguments{
\item{df1}{A data frame with the results of a functional enrichment analysis for the species 1 with an extra column "feature" with
the features to be compared}

\item{df2}{A data frame with the results of a functional enrichment analysis for the species 2 with an extra column "feature" with
the features to be compared}

\item{GOterm_field}{This is a string with the column name of the GO terms (e.g; "Functional_Category")}

\item{species1}{This is a string with the species name for the species 1 (e.g; "H. sapiens")}

\item{species2}{This is a string with the species name for the species 2 (e.g; "A. thaliana")}
}
\value{
This function will return a list with four slots: graphic, distance shared_GO_list,and unique_GO_list
}
\description{
compareGOspecies provides a simple workflow to compare  results of functional enrichment analysis
 for two species and a column which represent features to be compared.
 This function will extract the unique GO terms for two matrices matrices provided and will generate a presence-
 absence matrix where rows will represent a combination of categories and species(e.g H.sapies AID) and columns will be
 the GO terms analyzed.Further, this code will calculate Jaccard distances and will provide as outputs a principal
 coordinates analysis (PCoA), the jaccard distance matrix, the list of shared GO terms between species and finally,
 a list of the unique GO terms and the belongig to the species.
}
\examples{

#Loading example datasets
data(H_sapiens_compress)
data(A_thaliana_compress)
#Defining the column with the GO terms to be compared
GOterm_field <- "Functional_Category"
#Defining the species names
species1 <- "H. sapiens"
species2 <- "A. thaliana"
#Running function
x <- compareGOspecies(H_sapiens_compress,A_thaliana_compress,GOterm_field,species1,species2)
#Displaying PCoA results
x$graphic
# Checking shared GO terms between species
print(tapply(x$shared_GO_list$feature,x$shared_GO_list$feature,length))
# Checking unique GO terms for each species
print(tapply(x$unique_GO_list$species,x$unique_GO_list$species,length))
}
