% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comparison_example}
\alias{comparison_example}
\title{Functional enrichment analysis comparison between H. sapiens and A. thaliana}
\format{
An object of class \code{list} of length 4.
}
\source{
\url{https://data.mendeley.com/datasets/myyy2wxd59/1}
}
\usage{
comparison_example
}
\description{
This dataset is the results of running the compareGOspecies species and
 it is composed of four slots:
\describe{
  \item{graphics}{PCoA graphics}
  \item{distance}{numeric: Jaccard distance matrix }
  \item{shared_GO_list}{data.frame with shared GO terms between species}
  \item{unique_GO_list}{data.frame with unique GO terms and their belonging two each species}
}
}
\references{
Clavijo-Buriticá, Sosa, C.C., Mosquera, A.J. Álvarez, A., Medina, J. Quimbaya,
M.A. A systematic comparison of the molecular machinery associated with Cancer-Hallmarks
between plants and humans reveals Arabidopsis thaliana as a useful model to understand
specific carcinogenic events (to be submitted, Target journal: Plos Biology)
}
\keyword{datasets}
