% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOCompare-package.R
\docType{package}
\name{GOCompare-package}
\alias{GOCompare-package}
\alias{GOCompare}
\title{GOCompare: An R package to compare GO terms of gene lists (categories)  and their orthologs}
\description{
GOCompare is a an R package used to compare a GO terms list between
 two species
}
\details{
\tabular{ll}{
Package: \tab GOCompare\cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
Date: \tab 2022-11-07\cr
License: \tab GPL-3\cr
}
}
\keyword{package}
