% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_two_GOspecies.R
\name{graph_two_GOspecies}
\alias{graph_two_GOspecies}
\title{Comprehensive visual and graph comparison between two species and
 a series of categories}
\usage{
graph_two_GOspecies(
  x,
  species1,
  species2,
  GOterm_field,
  saveGraph = FALSE,
  option = 2,
  numCores = 2,
  outdir = NULL
)
}
\arguments{
\item{x}{is a list of running the comparegOspecies species.}

\item{species1}{This is a string with the species name for the species 1 (e.g; "H. sapiens")}

\item{species2}{This is a string with the species name for the species 2 (e.g; "A. thaliana")}

\item{GOterm_field}{This is a string with the column name of the GO terms (e.g; "Functional_Category")}

\item{saveGraph}{logical, if \code{TRUE} the function will allow save the graph in graphml format}

\item{option}{(values: 1 or 2). This option allows create either a graph
where nodes are GO terms and edges are features and species are edges arributes or
a graph where nodes are GO terms and edges are species belonging  (default value=2).}

\item{numCores}{numeric, Number of cores to use for the process (default value numCores=2)}

\item{outdir}{This parameter will allow save the graph file in a folder described here (e.g: "D:").This parameter only
works when saveGraph=TRUE}
}
\value{
This function will return a table representing an edge list
}
\description{
graph_two_GOspecies is a function to create undirected graphs
 to compare GO terms between two species using two options:
1.) Nodes are GO terms such as biological processes and the edges are features and species are edges attributes
2.) Nodes are GO terms such as biological processes and species status (e.g. A.thaliana, H. sapiens or shared) are edgess
}
\examples{

GOterm_field <- "Functional_Category"
data(comparison_example)
#Defining the species names
species1 <- "H. sapiens"
species2 <- "A. thaliana"
x_graph <- graph_two_GOspecies(x=comparison_example,
          species1=species1,
          species2=species2,
          GOterm_field=GOterm_field,
          numCores=2,
          saveGraph = FALSE,
          option=1,
          outdir = NULL)
head(x_graph)
}
