% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_GOspecies.R
\name{graphGOspecies}
\alias{graphGOspecies}
\title{network representation in format graphml for the results of functional
 enrichment analysis for one species}
\usage{
graphGOspecies(
  df,
  GOterm_field,
  option = 2,
  numCores = 2,
  saveGraph = FALSE,
  outdir = NULL
)
}
\arguments{
\item{df}{A data frame with the results of a functional enrichment analysis for
a species with an extra column "feature" with the features to be compared}

\item{GOterm_field}{This is a string with the column name of the GO terms (e.g; "Functional.Category")}

\item{option}{(values: 1 or 2). This option allows create either a graph where nodes are GO terms and edges are features or
a graph where nodes are features and edges are GO terms (default value=2).}

\item{numCores}{numeric, Number of cores to use for the process (default value numCores=2)}

\item{saveGraph}{logical, if \code{TRUE} the function will allow save the graph in graphml format}

\item{outdir}{This parameter will allow save the graph file in a folder described here (e.g: "D:").This parameter only
works when saveGraph=TRUE}
}
\value{
This function will return a table representing an edge list
}
\description{
graphGOspecies is a function to create undirected graphs using two options:
1.) Nodes are GO terms such as biological processes and the edges are features. Edges weights are calculated but this function
requires a long time to finish.
2.) Nodes are features, the edges are the number of GO terms such as biological processes and they are represented
such as weights and collapsed in a column called FEATURES
}
\examples{

#Loading example datasets
data(H_sapiens_compress)
#Defining the column with the GO terms to be compared
GOterm_field <- "Functional_Category"
#Running function
x <- graphGOspecies(df=H_sapiens_compress,
                     GOterm_field=GOterm_field,
                     option = 2,
                     numCores=2,
                     saveGraph=FALSE,
                     outdir = NULL)
#Displaying results
head(x)
}
